Don't delete last archived state

+++ beacon-chain/state/stategen/migrate.go
@@ -74,7 +74,8 @@ func (s *State) MigrateToCold(ctx context.Context, finalizedSlot uint64, finaliz
 			// Do not delete the current finalized state in case user wants to
 			// switch back to old state service, deleting the recent finalized state
 			// could cause issue switching back.
-			if s.beaconDB.HasState(ctx, r) && r != finalizedRoot {
+			lastArchivedIndexRoot := s.beaconDB.LastArchivedIndexRoot(ctx)
+			if s.beaconDB.HasState(ctx, r) && r == lastArchivedIndexRoot && r != finalizedRoot {
 				if err := s.beaconDB.DeleteState(ctx, r); err != nil {
 					// For whatever reason if node is unable to delete a state due to
 					// state is finalized, it is more reasonable to continue than to exit.
